Now including lang-specific Bash configs again
This time it was attempted to make it so that the lang-specific configs
only load if their configuration is relevant (ie - their runtime is
available on the path for example).

+++ install.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
 
-ln -s ~/dotfiles/.bashrc ~/.bashrc
+#-------------------------------#
+# SYMLINK - .bashrc             #
+#-------------------------------#
+if [[ ! -L ~/.bashrc ]]; then
+  ln -s ~/dotfiles/.bashrc ~/.bashrc
+fi
+
